requirements.txt
Flask
python-dotenv
playwright
beautifulsoup4
tqdm
openpyxl
gunicorn